feat(accounting): delineate value and quantity concepts wrt Amount
We have added two new functions `amountFromValue` and
`valueFromAmount` in addition to existing `amountFromQuantity` and
`quantityFromAmount` functions.

The former refer to *value* as in net effect on the net-worth of the
entity, whereby the latter are rather conventional as in net effect on
the account as per 'AccountKind'.

Functions are documented and examples are given to explain the
differences.